updated db

+++ bamazon.sql
@@ -10,7 +10,7 @@ CREATE TABLE products (
     id INTEGER(10) NOT NULL AUTO_INCREMENT,
     product_name VARCHAR(50) NOT NULL,
     department_name VARCHAR(50) NOT NULL,
-    price INTEGER(10),
+    price DECIMAL(5, 2),
     stock_quantity INTEGER(10),
     PRIMARY KEY (id)
 );
